+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Runtime/SceneReference.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Linq;
 using System.Runtime.Serialization;
 using System.Xml;
 using System.Xml.Schema;
@@ -28,14 +27,27 @@ namespace Eflatun.SceneReference
         /// </summary>
         /// <param name="scenePath">Path of the scene to reference.</param>
         /// <returns>A new <see cref="SceneReference"/>.</returns>
-        /// <remarks>
-        /// This factory method does NOT validate the given path. If it is invalid, then the created
-        /// <see cref="SceneReference"/> will also be invalid.
-        /// </remarks>
+        /// <exception cref="SceneReferenceCreationException">Throws if the given path is null or whitespace.</exception>
+        /// <exception cref="SceneReferenceCreationException">Throws if the given path is not found in the Scene Path to GUID map.</exception>
         public static SceneReference FromScenePath(string scenePath)
         {
-            var guid = SceneGuidToPathMapProvider.SceneGuidToPathMap.SingleOrDefault(x => x.Value == scenePath).Key;
-            return new SceneReference(guid);
+            if (string.IsNullOrWhiteSpace(scenePath))
+            {
+                throw new SceneReferenceCreationException(
+                    $"Given path is null or whitespace. Path: '{scenePath}'" +
+                    "\nTo fix this, make sure you provide the path of a valid scene.");
+            }
+
+            if (!SceneGuidToPathMapProvider.ScenePathToGuidMap.TryGetValue(scenePath, out var sceneGuid))
+            {
+                throw new SceneReferenceCreationException(
+                    $"Given path is not found in the scene GUID to path map. Path: '{scenePath}'"
+                    + "\nThis can happen for these reasons:"
+                    + "\n1. The asset at the given path either doesn't exist or is not a scene. To fix this, make sure you provide the path of a valid scene."
+                    + "\n2. The scene GUID to path map is outdated. To fix this, you can either manually run the generator, or enable generation triggers. It is highly recommended to keep all the generation triggers enabled.");
+            }
+
+            return new SceneReference(sceneGuid);
         }
 
         // GUID hex of an invalid asset contains all zeros. A GUID hex has 32 chars.
@@ -45,8 +57,9 @@ namespace Eflatun.SceneReference
         [SerializeField] internal string sceneAssetGuidHex;
 
         /// <summary>
-        /// Creates a new <see cref="SceneReference"/> which is empty, and subsequently invalid.
+        /// Creates a new empty <see cref="SceneReference"/>.
         /// </summary>
+        /// <remarks>This constructor never throws.</remarks>
         public SceneReference()
         {
             // This parameterless constructor is required for the XML serialization.
@@ -60,18 +73,40 @@ namespace Eflatun.SceneReference
         /// Creates a new <see cref="SceneReference"/> which references the scene that has the given GUID.
         /// </summary>
         /// <param name="sceneAssetGuidHex">GUID of the scene to reference.</param>
-        /// <remarks>
-        /// This constructor does NOT validate the given GUID. If it is invalid, then the created
-        /// <see cref="SceneReference"/> will also be invalid.
-        /// </remarks>
+        /// <exception cref="SceneReferenceCreationException">Throws if the given GUID is null or empty.</exception>
+        /// <exception cref="SceneReferenceCreationException">Throws if the given GUID is not found in the Scene GUID to Path map.</exception>
+        /// <exception cref="SceneReferenceCreationException">(Editor-only) Throws if the asset is not found at the path that the GUID corresponds to.</exception>
         public SceneReference(string sceneAssetGuidHex)
         {
+            if (string.IsNullOrWhiteSpace(sceneAssetGuidHex))
+            {
+                throw new SceneReferenceCreationException(
+                    $"Given GUID is null or whitespace. GUID: '{sceneAssetGuidHex}'." +
+                    "\nTo fix this, make sure you provide the GUID of a valid scene.");
+            }
+
+            if (!SceneGuidToPathMapProvider.SceneGuidToPathMap.TryGetValue(sceneAssetGuidHex, out var scenePath))
+            {
+                throw new SceneReferenceCreationException(
+                    $"Given GUID is not found in the scene GUID to path map. GUID: '{sceneAssetGuidHex}'"
+                    + "\nThis can happen for these reasons:"
+                    + "\n1. The asset with the given GUID either doesn't exist or is not a scene. To fix this, make sure you provide the GUID of a valid scene."
+                    + "\n2. The scene GUID to path map is outdated. To fix this, you can either manually run the generator, or enable generation triggers. It is highly recommended to keep all the generation triggers enabled.");
+            }
+
             this.sceneAssetGuidHex = sceneAssetGuidHex;
 
 #if UNITY_EDITOR
-            sceneAsset = SceneGuidToPathMapProvider.SceneGuidToPathMap.TryGetValue(sceneAssetGuidHex, out var scenePath)
-                ? AssetDatabase.LoadAssetAtPath<UnityEngine.Object>(scenePath)
-                : null;
+            var foundAsset = AssetDatabase.LoadAssetAtPath<UnityEngine.Object>(scenePath);
+
+            if (!foundAsset)
+            {
+                throw new SceneReferenceCreationException(
+                    $"The given GUID was found in the map, but the scene asset at the corresponding path could not be loaded. Path: '{scenePath}'."
+                    + "\nThis can happen due to an outdated scene GUID to path map retaining scene assets that no longer exist. To fix this, you can either manually run the generator, or enable generation triggers. It is highly recommended to keep all the generation triggers enabled.");
+            }
+
+            sceneAsset = foundAsset;
 #endif // UNITY_EDITOR
         }
 
@@ -80,17 +115,32 @@ namespace Eflatun.SceneReference
         /// Creates a new <see cref="SceneReference"/> which references the given scene asset.
         /// </summary>
         /// <param name="sceneAsset">The asset of the scene to reference.</param>
-        /// <remarks>
-        /// This constructor is for editor-use only. Do NOT use it in runtime code.<p/>
-        /// This constructor does NOT validate the given asset. If it is invalid, then the created
-        /// <see cref="SceneReference"/> will also be invalid.
-        /// </remarks>
+        /// <exception cref="SceneReferenceCreationException">Throws if the given asset is null.</exception>
+        /// <exception cref="SceneReferenceCreationException">Throws if the GUID of the given asset cannot be retrieved.</exception>
+        /// <exception cref="SceneReferenceCreationException">Throws if the Scene GUID to Path map does not contain the GUID of the given asset.</exception>
+        /// <remarks>This constructor is for editor-use only. Do NOT use it in runtime code.</remarks>
         public SceneReference(UnityEngine.Object sceneAsset)
         {
-            sceneAssetGuidHex = AssetDatabase.TryGetGUIDAndLocalFileIdentifier(sceneAsset, out var guid, out long _)
-                ? guid
-                : AllZeroGuidHex;
+            if (!sceneAsset)
+            {
+                throw new SceneReferenceCreationException("Given scene asset is null. To fix this, make sure you provide a valid scene asset.");
+            }
 
+            if (!AssetDatabase.TryGetGUIDAndLocalFileIdentifier(sceneAsset, out var guid, out long _))
+            {
+                throw new SceneReferenceCreationException("Could not retrieve the GUID of the given scene asset. This usually indicates an invalid asset. To fix this, make sure you provide a valid scene asset.");
+            }
+
+            if (!SceneGuidToPathMapProvider.SceneGuidToPathMap.ContainsKey(guid))
+            {
+                throw new SceneReferenceCreationException(
+                    $"The GUID of the given scene asset is not found in the scene GUID to path map. GUID: '{guid}'"
+                    + "\nThis can happen for these reasons:"
+                    + "\n1. Given asset either doesn't exist or is not a scene. To fix this, make sure you provide a valid scene asset."
+                    + "\n2. The scene GUID to path map is outdated. To fix this, you can either manually run the generator, or enable generation triggers. It is highly recommended to keep all the generation triggers enabled.");
+            }
+
+            sceneAssetGuidHex = guid;
             this.sceneAsset = sceneAsset;
         }
 #endif // UNITY_EDITOR